+++ want/tests/test11/run.sh
@@ -209,8 +209,8 @@ if [ "$CONDUCTOR" = ".TRUE." ] ; then
    $WANT_BIN/conductor.x < $TEST_HOME/conductor.in > $TEST_HOME/conductor.out
    if [ ! -e CRASH ] ; then 
       echo "$ECHO_T done" 
-      mv dos.dat $TEST_HOME/dos.dat
-      mv cond.dat $TEST_HOME/cond.dat
+      test -e dos.dat && mv dos.dat $TEST_HOME/dos.dat
+      test -e cond.dat && mv cond.dat $TEST_HOME/cond.dat
    else
       echo "$ECHO_T problems found" ; cat CRASH ; exit 1
    fi
@@ -225,8 +225,8 @@ if [ "$CONDUCTOR_BULK" = ".TRUE." ] ; then
    $WANT_BIN/conductor.x < $TEST_HOME/conductor_bulk.in > $TEST_HOME/conductor_bulk.out
    if [ ! -e CRASH ] ; then 
       echo "$ECHO_T done" 
-      mv dos.dat $TEST_HOME/dos_bulk.dat
-      mv cond.dat $TEST_HOME/cond_bulk.dat
+      test -e dos.dat && mv dos.dat $TEST_HOME/dos_bulk.dat
+      test -e cond.dat && mv cond.dat $TEST_HOME/cond_bulk.dat
    else
       echo "$ECHO_T problems found" ; cat CRASH ; exit 1
    fi
